Registers in Computer Architecture (C.A)
This repository provides an in-depth exploration of **Registers** in computer architecture. Registers are critical components of the CPU, designed to store and transfer data and instructions at extremely high speeds. This project discusses their types, functions, and relevance in computer systems, along with their interaction with key CPU units like the ALU and CU.